Create README.md
Currently my firmware produces a power outage code after about an hour of printing.. This leaves a message continue print or abandon print . when I push the button the print will continue the print continues but there is a gap in the print. I've sat and watched the error reproduce and there is no actual power outage that occurs.